requirements.txt
gradio
ultralytics
supervision==0.15.0
requests==2.25.1
gofile==0.2.0
pyngrok==7.0.0
ffmpeg==1.4
Flask==2.1.3
python-dotenv==1.0.0
